refactor: inline TopicsMasterDetail in topics/page.tsx
